+++ src/typopo.js
@@ -1,10 +1,10 @@
 /*!
- * Typopo 1.0.0
+ * Typopo 1.0.1
  *
  * Copyright 2015-16 Braňo Šandala
  * Released under the MIT license
  *
- * Date: 2016-03-28
+ * Date: 2016-04-10
  */
 
 (function(){
@@ -19,6 +19,8 @@ var essential_set = {
     "\\-\\+": "±",
 };
 
+
+
 var lowercase_chars_en_sk_cz_rue = "a-záäčďéěíĺľňóôöőŕřšťúüűůýžабвгґдезіийклмнопрстуфъыьцчжшїщёєюя";
 var uppercase_chars_en_sk_cz_rue = "A-ZÁÄČĎÉĚÍĹĽŇÓÔÖŐŔŘŠŤÚÜŰŮÝŽАБВГҐДЕЗІИЙКЛМНОПРСТУФЪЫЬЦЧЖШЇЩЁЄЮЯ";
 var single_quote_adepts = "‚'‘’‛";
@@ -32,13 +34,19 @@ function replace_symbols(string, replacement_set) {
     return string;
 }
 
+
+
 function replace_periods_with_ellipsis(string) {
     return string.replace(/\.{2,}/g, "…");
 }
 
+
+
 function remove_multiple_spaces(string) {
     return string.replace(/ {2,}/g, " ");
 }
+
+
 
 function correct_double_quotes(string, language) {
     var pattern = "(" + double_quote_adepts + ")(.*?)(" + double_quote_adepts + ")";
@@ -62,13 +70,19 @@ function swap_quotes_and_punctuation(string) {
     // return string.replace(/([“”])([\.,!?])/g, "$2$1");
 }
 
+
+
 function correct_multiple_sign(string) {
     return remove_multiple_spaces(string.replace(/([1-9]+[ ]{0,1}[a-wz]*)([ ]{0,1}[x|×][ ]{0,1})([1-9]+[ ]{0,1}[a-wz]*)/g, "$1 × $3"));
 }
 
+
+
 function replace_hyphen_with_dash(string) {
     return string.replace(/( [-|–] )/g, " — ");
 }
+
+
 
 function replace_dash_with_hyphen(string){
     var pattern = "(["+ lowercase_chars_en_sk_cz_rue +"])([–—])(["+ lowercase_chars_en_sk_cz_rue +"])";
@@ -79,13 +93,13 @@ function replace_dash_with_hyphen(string){
 
 
 function remove_space_before_punctuation(string) {
-    return string.replace(/([ ])([\,\.\!\?\:\;\)])/g, "$2");
+    return string.replace(/([ ])([\,\.\!\?\:\;\)\]\}])/g, "$2");
 }
 
 
 
 function remove_space_after_punctuation(string) {
-    return string.replace(/([\(])([ ])/g, "$1");
+    return string.replace(/([\(\[\{])([ ])/g, "$1");
 }
 
 
@@ -143,6 +157,8 @@ function remove_spaces_at_paragraph_beginning(string) {
     return lines.join("\n");
 }
 
+
+
 function start_sentence_w_capital_letter(string) {
     // Correct sentence for the first initial sentence
     var lines = string.match(/[^\r\n]+/g);
@@ -170,6 +186,8 @@ function start_sentence_w_capital_letter(string) {
 
     return string;
 }
+
+
 
 function correct_accidental_uppercase(string) {
     // var pattern = "[a-z]+[a-zA-Z]+[A-Z]+|[A-Z]+[a-zA-Z]+[a-z]+|[a-z]+[a-zA-Z]+[a-z]+";
@@ -237,6 +255,7 @@ function identify_common_apostrophes(string) {
 }
 
 
+
 /*
     Identifies single quotes that may be used within double quotes and replace them with identifiers
 
@@ -292,6 +311,36 @@ function identify_residual_apostrophes(string) {
 
 
 /*
+    Corrects improper spacing around aposiopesis
+    Due to the fact that ellipsis character is being used both as an ellipsis and as an aposiopesis, we are able to auto-correct these scenarios:
+    [1] remove space before aposiopesis, that is ending a sentence
+    [2] remove space when aposiopesis is used at the beginning of the sentence
+
+    @param {string} string — input text for identification
+    @returns {string} — output with identified apostrophes
+*/
+function correct_spaces_around_aposiopesis(string) {
+    /*[1] catching ending sentences in the middle of the paragraph*/
+    var pattern = "([" + lowercase_chars_en_sk_cz_rue + "])( )(… [" + uppercase_chars_en_sk_cz_rue + "])";
+    var re = new RegExp(pattern, "g");
+    string = string.replace(re, "$1$3");
+
+    /*[1] catching ending sentences at the end of the paragraph*/
+    pattern = "([" + lowercase_chars_en_sk_cz_rue +"])( )(…)(?![ " + lowercase_chars_en_sk_cz_rue + uppercase_chars_en_sk_cz_rue + "])";
+    re = new RegExp(pattern, "g");
+    string = string.replace(re, "$1$3");
+
+    /*[2]*/
+    pattern = "([\.\?\!] …)( )([" + lowercase_chars_en_sk_cz_rue +"])";
+    re = new RegExp(pattern, "g");
+    string = string.replace(re, "$1$3");
+
+    return string;
+}
+
+
+
+/*
     Identifies intended use of apostrophes and single quotes and adjusts them accordingly.
 
     Assumptions (meaning that author will use quotes in the following manner):
@@ -338,6 +387,15 @@ function correct_single_quotes_and_apostrophes(string, language) {
 }
 
 
+/*
+    Sets proper spacing, when ellipsis used used around commas
+
+    @param {string} string — input text for identification
+    @returns {string} — corrected output
+*/
+function space_ellispsis_around_commas(string) {
+    return string.replace(/, ?… ?,/g, ", …,");
+}
 
 // supported languages: en, sk, cs, rue
 function correct_typos(string, language) {
@@ -360,6 +418,8 @@ function correct_typos(string, language) {
     string = add_space_after_punctuation(string);
     string = remove_spaces_at_paragraph_beginning(string);
     string = replace_with_nbsp(string);
+    string = space_ellispsis_around_commas(string);
+    string = correct_spaces_around_aposiopesis(string);
 
     string = correct_multiple_sign(string);
     string = replace_hyphen_with_dash(string);